Clean up code style and remove unwanted comments in the tests directories.
